Experiments with extending this project for generating Istio certs.
Main issue is that K8S CA can't issue URL certs, and it doesn't appear
to be able to generate intermediary CA either.

So the only way to use K8S CA for Istio would be to use DNS-formated
form of the spiffee, with some changes in pilot to use that identity.